backend/requirements.txt
# DeepSearch Backend Dependencies
# 版本: v1.0

# ============================================
# Web Framework
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ============================================
# Database
# ============================================
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# ============================================
# Validation
# ============================================
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# ============================================
# Task Queue
# ============================================
celery>=5.3.6
redis>=5.0.1
flower>=2.0.1

# ============================================
# Search Engine
# ============================================
meilisearch>=0.31.0

# ============================================
# Document Parsing
# ============================================
kreuzberg>=3.22.0
python-magic>=0.4.27
Pillow>=10.2.0

# ============================================
# HTTP Client
# ============================================
httpx>=0.27.0
aiohttp>=3.9.3

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.1
orjson>=3.9.13
tenacity>=8.2.3
aiofiles>=23.2.1

# ============================================
# Logging
# ============================================
structlog>=24.1.0

# ============================================
# Testing
# ============================================
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0

# ============================================
# Development
# ============================================
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
